Implement routing and forms for user and subject management in the frontend
It has been dispayed on form for each CRUD endpoint to be able to test endpoints. The function onSubmit is right now by default, but are going to be implemented in a future issue some hooks to manage the real requests. The objective of this thask is to create a simple control pannel based in forms.

There are custom componets usig the base from the component shadcn/ui form component.